[TECH-CHALLENGE-FASE-2] - adicionando CRUD para restaurante e item cardapio

+++ src/main/java/com/raidstack/restmanager/entity/Restaurante.java
@@ -15,7 +15,7 @@ import java.sql.Time;
 @Table("restaurantes")
 public class Restaurante {
 
-    private int id;;
+    private Long id;;
     private String nome;
     private String endereco;
     private String tipo_cozinha;
